requirements.txt
torch>=2.0.0
safetensors
Pillow
numpy
modelscope
huggingface_hub